♻️ refactor(dockerfile): refactor Dockerfile image layers
The image layers for the Dockerfile previously didn't make sense, so I refactored them. They may
still need some work, I think they are too complicated at the moment. I am trying to make it so that
I can use one layer for different dev environments, but I'm wondering if that is not the point of
using different layers in Dockerfiles.